+++ src/library/consoleObj.cpp
@@ -83,4 +83,11 @@ void consoleObj::init(stack *st)
             (*args[0])(args[0], object::arrayType(args.begin() + 1, args.end()), st);
         }));
     });
+    addFunctionL(consoleObj, "setInput"_n, {
+        argsGuard<std::string, std::nullptr_t>(args);
+       console::setInput(args[0]->getConverted<std::string>(), [&args, &st]() {
+            (*args[1])(args[1], object::arrayType(args.begin() + 2, args.end()), st);
+        });
+        return thisObj;
+    });
 }